Add option for `expiredBuffer`
When checking if an access token has expired, there is a common race
condition where using a near-expiring token may actually be expired by
the time the API server checks it. This is due to things like round-trip
latency, clock skew, processing time, etc.

One solution is to allow clients to configure a buffer so that a token
will be reported as "expired" before it actually is. This allows some
leeway to account for various latencies.

Fixes #131