Documentation Fix - Passing multiple scopes
The documentation indicates that multiple scopes may be passed as an array of strings.  However, this results in an incorrect payload due to the way `stringify` handles the url encoding of arrays.  If an array is passed for `scope` the resulting URL contains multiple `scope` parameters. e.g.
```scope=scope-1&scope=scope-2```
As per [Sec 3.3 of the OAuth spec](https://tools.ietf.org/html/draft-ietf-oauth-v2-31#section-3.3), scope should be passed as a space-delimited string.  Therefore, when multiple scopes are desired, the `scope` parameter should contain a space delimited string of the desired scopes. ex.
```scope: 'scope-1 scope-2```